[tests] Use a custom wrapper script to check the return code of the MPI test binaries
ULFM's mpirun does not pass through the return code of the binary it
executes. We therefore resorted to grepping the output of the test
binary for PASSED/FAILED.

This patch introduces a wrapper script around the MPI tests that allows
us to actually check their return values. A special return value (42) is
used to signal that a testcase's rank terminated to simulate a failure.